requirements.txt
# -*- coding: utf-8 -*-
"""
Created on Wed Jul 16 13:51:26 2025

@author: bongi
"""

openai>=1.0.0
python-dotenv>=1.0.0
pymupdf>=1.22.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=9.0.0
